requirements.txt
flask
python-dotenv
Session
sqlalchemy
psycopg2